Use more reliable approach for Zonar bus updates
Instead of relying on a sliding time window, since we know the time
of the most recent bus position, we can explicitly ask for all events
since that time. In the case we have a transient Zonar failure, we'll
still collect the events we want. There is a 5-minute cutoff to ensure
that a longer outage won't result in a massive request later.