Split translations into separate files
This will be necessary for automated upload to a translation service.
For now, to avoid having to duplicate all of the "locale names" across
every locale, this switches ember-i18n to a not-yet-merged Git branch
that allows translating into locales other than the current one.

Also, removed a "Cancel" link that was supposed to be removed earlier.