Don't load translations for the main landing page
We are currently not offering the "main site" in other languages, so
it does not make sense to wait for an extra server request to get the
translations before the page loads. This does necessitate literalizing
the text for the Terms of Service and Privacy Policy links, but this is
acceptable as the documents themselves are only available in English.