Refactor "current district" handling and add error page
Previously if someone went to a nonexistent schoolbot.io subdomain,
it would unhelpfully render the marketing page for the root domain
(but not redirect you to the root domain URL), instead of telling you
that something went wrong. Now we distinguish these cases, and have a
friendly error page.

As a related change, the "current district" is no longer retrieved
using an API call, but is instead pre-serialized as part of the
initial client HTML request, and pushed into the store on Ember
startup. Since the root domain is all static English copy and does
not need to load translations, this gets us down to zero AJAX
requests on that page.