Fix errors being suppressed in the development environment
This is tricky because A) PhantomJS seems to react in an extremely
glitchy manner when re-throwing an error from the Ember error handler,
and B) we don't want thrown Ember errors to fail feature specs that are
testing the error handling itself. This commit uses separate methods to
(awkwardly) work around each of these, but there should be a more clean
way to do this.