Use session cookie instead of custom session store
This has the same effect as the previous session store (clearing data
when the tab is closed), but without all the custom code, and with
the future ability to set an expiration for the cookie if the user picks a "Remember me" option.

This also folds storage of the preferred locale into the ESA session store, since it operates just fine without an authenticated user.