Set correct domain length for URL generation
By default Rails assumes the length of your "root domain" is always 2,
so generating a URL with the `subdomain` option could in some cases
replace the leftmost part of the APPLICATION_HOST rather than adding
onto it.

This also removes the separate ActionMailer `default_url_options`
configuration, since it turns out to use the defaults defined in the
router anyway.