Adjust RACK_TIMEOUT to Suspenders default value
Currently there are very occasional timeouts happening on production,
presumably due to fluctuating load. When these occur in a certain part
of ActiveRecord, it can be left in an inconsistent state, causing
subsequent queries to fail with a "prepared statement already exists"
error. This makes those cascading errors less likely.