Allow districts to be set as "inactive"
Inactive districts basically do not exist for all purposes – users
will see a not-found page when accessing the subdomain, Zonar updates
are not pulled, and the assignments API returns 401.

This allows us to leave districts "inactive" on staging unless we
currently need them, so both staging and production will not be making
redundant requests for the same Zonar data.