Fix deleted users not being deleted in Intercom
Since passing destroyed models through GID serialization is not
possible, we need to pass just an ID, which means checking the class
of the argument passed to the job is no longer sufficient to determine
the action that should be done. Now the action to perform is passed as
an additional argument to the job.