Replace Poltergeist with capybara-webkit
Since we originally switched to Poltergeist, capybara-webkit has
advanced to the point that we can use it without errors. In the
meantime, frontend dependency upgrades have become progressively
harder to do without destabilizing PhantomJS, specifically around the
specs that require throwing JavaScript errors.

Unfortunately this does add about 1:30 to the CircleCI test runtime,
due to the inability to cache apt dependencies. However, it might be
possible to pre-download these packages to a cached directory, and use
dpkg to install them from there.